+++ src/api/supabase/queries/user_queries.ts
@@ -9,6 +9,8 @@ import {
 } from '@supabase/supabase-js';
 import { User } from '../../../schema/schema';
 
+import { Product } from '../../../schema/schema';
+
 // Replace these with your Supabase project URL and API key
 const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
 const supabaseApiKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY;
@@ -53,88 +55,110 @@ export async function fetchUserByUUID(uuid: string) {
   }
 }
 
-export async function fetchFavoriteItems(
-  userId: string,
-): Promise<Record<string, number>> {
-  // Fetch fav_items for the specified user
-  const { data, error } = await supabase
-    .from('profiles')
-    .select('fav_items')
-    .eq('user_id', userId)
-    .single();
+export async function getUserInfo(product: Product, isFav: Boolean) {
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
 
-  if (error) {
-    throw new Error(
-      `An error occurred when trying to fetch favorite items: ${error.message}`,
-    );
-  } else if (!data) {
-    throw new Error('No user found with the specified user_id.');
-  }
+  if (user != null) {
+    const { data, error } = await supabase
+      .from('profiles')
+      .select()
+      .eq('user_id', user.id);
 
-  return data.fav_items;
-}
+    if (data != null) {
+      const CurrUserFavoriteItems = data[0].fav_items;
 
-export async function addToFavorites(userId: string, productId: string) {
-  // First, fetch the current fav_items for the user
-  const { data, error } = await supabase
-    .from('profiles')
-    .select('fav_items')
-    .eq('user_id', userId)
-    .single();
+      if (isFav) {
+        CurrUserFavoriteItems[product.product_id] = 1;
+      } else {
+        delete CurrUserFavoriteItems[product.product_id];
+      }
 
-  if (error) {
-    throw new Error(
-      `An error occurred when trying to fetch the user's favorite items: ${error.message}`,
-    );
-  }
-
-  const currentFavItems = data?.fav_items || {};
-
-  // Add the product to fav_items or update its quantity
-  currentFavItems[productId] = currentFavItems[productId] || 0;
-
-  // Now update the user's fav_items in the database
-  const updateResponse = await supabase
-    .from('profiles')
-    .update({ fav_items: currentFavItems })
-    .eq('user_id', userId);
-
-  if (updateResponse.error) {
-    throw new Error(
-      `An error occurred when trying to update the user's favorite items: ${updateResponse.error.message}`,
-    );
+      const { error } = await supabase
+        .from('profiles')
+        .update({ fav_items: CurrUserFavoriteItems })
+        .eq('user_id', user.id);
+    }
   }
 }
 
-// Function to remove a product from fav_items
-export async function removeFromFavorites(userId: string, productId: string) {
-  // First, fetch the current fav_items for the user
-  const { data, error } = await supabase
-    .from('profiles')
-    .select('fav_items')
-    .eq('user_id', userId)
-    .single();
+export async function arrayOfFavorites() {
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
 
-  if (error) {
-    throw new Error(
-      `An error occurred when trying to fetch the user's favorite items: ${error.message}`,
-    );
+  if (user != null) {
+    const { data, error } = await supabase
+      .from('profiles')
+      .select()
+      .eq('user_id', user.id)
+      .single();
+
+    const CurrUserFavoriteItems = data.fav_items;
+
+    const Favkey = Object.keys(CurrUserFavoriteItems);
+
+    var FavArray = [];
+
+    for (let i = 0; i < Favkey.length; i++) {
+      let key = Favkey[i];
+      const { data, error } = await supabase
+        .from('product')
+        .select()
+        .eq('product_id', key)
+        .single();
+
+      if (data != undefined) {
+        FavArray.push(data);
+      }
+    }
+
+    return FavArray;
   }
+}
 
-  const currentFavItems = data?.fav_items || {};
+export async function totalNumberOfItemsInCart() {
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
 
-  // Remove the product from fav_items
-  delete currentFavItems[productId];
-  console.log(currentFavItems);
-  // Now update the user's fav_items in the database
-  const updateResponse = await supabase
-    .from('profiles')
-    .update({ fav_items: currentFavItems })
-    .eq('user_id', userId);
+  if (user != null) {
+    const { data, error } = await supabase
+      .from('profiles')
+      .select()
+      .eq('user_id', user.id)
+      .single();
 
-  if (updateResponse.error) {
-    throw new Error(
-      `An error occurred when trying to update the user's favorite items: ${updateResponse.error.message}`,
-    );
+    if (data != null) {
+      const CurrUserCart = data.cart;
+      console.log(CurrUserCart);
+      if (CurrUserCart == null || CurrUserCart.length == 0) {
+        return 0;
+      } else {
+        const itemNumb = Object.values(CurrUserCart) as number[];
+
+        let sum = 0;
+
+        for (let i = 0; i < itemNumb.length; i++) {
+          sum = sum + itemNumb[i];
+        }
+
+        return sum;
+      }
+    }
   }
+}
+
+export async function getProduct() {
+  const { data } = await supabase.from('product').select('*');
+  return data;
+}
+
+export async function filterProduct(productType: String) {
+  const { data } = await supabase
+    .from('product')
+    .select('*')
+    .eq('category', productType);
+  return data;
 }
