+++ src/components/NavBar.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import Link from 'next/link';
 import React, { useEffect, useState } from 'react';
 
-import { totalNumberOfItemsInCart } from '../app/storefront/helperFunction';
+import { totalNumberOfItemsInCart } from '../api/supabase/queries/user_queries';
 
 import {
   NavBarComp,
